backend/user-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic[email]
email-validator
bcrypt
